* Fix codacy issues. * Move helm charts from k8s directory to root directory.
