+++ charts/vineyard-operator/templates/deployment.yaml
