+++ charts/vineyard-operator/templates/globalobject-crd.yaml
