+++ charts/vineyard-operator/templates/leader-election-rbac.yaml
