+++ charts/vineyard-operator/templates/manager-rbac.yaml
