+++ charts/vineyard-operator/templates/serving-cert.yaml
