+++ charts/vineyard-operator/templates/sidecar-crd.yaml
