+++ k8s/apis/k8s/v1alpha1/sidecar_webhook.go
@@ -17,7 +17,7 @@ limitations under the License.
 package v1alpha1
 
 import (
-	log "github.com/v6d-io/v6d/k8s/internal/log"
+	log "github.com/v6d-io/v6d/k8s/pkg/internal/log"
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/webhook"
