+++ k8s/pkg/internal/log/log.go
@@ -13,6 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+// Package log contains the global logger for the vineyard operator.
 package log
 
 import (
