+++ k8s/test/e2e/autogenerated-helm-chart/e2e.yaml
@@ -33,7 +33,7 @@ setup:
     - name: use helm chart to install the operator
       command: |
         kind load docker-image vineyardcloudnative/vineyard-operator:latest
-        cd k8s
+
         helm install vineyard-operator charts/vineyard-operator \
             --namespace vineyard-system \
             --create-namespace
